refine inference scripts, update readme
