--- example_diffold/1ake_dataset_monomer.json
@@ -1,10 +1,0 @@
-{
-    "1ake": {
-        "id": "1ake_A",
-        "initial_t": 1.0,
-        "num_replica": 3,
-        "num_steps": 10,
-        "inf_steps": 30,
-        "save_trajectory": true
-    }
-}
