--- example_diffold/2dn1_2dn2_from_pdb.json
@@ -1,22 +1,0 @@
-{
-    "2dn1": {
-        "pdb": "2dn1",
-        "residue_idx": "1-140/142-281/283-427/429-573",
-        "is_monomer": false,
-        "initial_t": 0.1,
-        "num_replica": 3,
-        "num_steps": 10,
-        "inf_steps": 30,
-        "save_trajectory": true
-    },
-    "2dn2": {
-        "pdb": "2dn2",
-        "residue_idx": "1-140/142-281/283-427/429-573",
-        "is_monomer": false,
-        "initial_t": 0.1,
-        "num_replica": 3,
-        "num_steps": 10,
-        "inf_steps": 30,
-        "save_trajectory": true
-    }
-}
