+++ inference_scripts_ufconf/run_ufconf_interpolate.py
@@ -5,19 +5,16 @@ import json
 import torch
 import time
 from typing import *
-import numpy as np
 import tqdm
 
 from absl import logging
 logging.set_verbosity("info")
-from ufconf.dataset import DiffoldDataset
+from unifold.dataset import process
 from ufconf.diffusion.diffuser import Diffuser
 from ufconf.diffuse_ops import diffuse_inputs, make_noisy_quats, rbf_kernel
-from unifold.data.protein import to_pdb
 import copy
 import random
 import ufconf.utils as utils
-from unifold.dataset import process
 
 max_retries = 3  # Maximum number of retries
 retry_delay = 3  # Delay between retries in seconds
@@ -50,94 +47,6 @@ def setup_directories(args, job_name_list, job_name, Job):
     os.makedirs(dir_feat_name, exist_ok=True)
 
     return output_dir, output_traj_dir, dir_feat_name
-
-def save_config_json(args, checkpoint_path, output_dir, Job):
-    """
-    Saves the job configuration to a JSON file in the output directory.
-
-    Args:
-        args: Parsed command-line arguments.
-        checkpoint_path (str): Path to the model checkpoint.
-        output_dir (str): Directory where the output is to be saved.
-        Job (dict): Configuration dictionary for the current job.
-    """
-    config_data = {
-        "timestamp": time.strftime("%Y-%m-%d;%H:%M:%S", time.localtime()),
-        "diffusion_t": Job["initial_t"],
-        "checkpoint": checkpoint_path
-    }
-
-    # Define the file path
-    config_file_path = os.path.join(output_dir, "config.json")
-
-    # Writing the configuration data to the JSON file
-    with open(config_file_path, 'w') as config_file:
-        json.dump(config_data, config_file, indent=2)
-        
-def prepare_features(feat, Job):
-    """
-    Prepares features based on the job configuration.
-
-    Args:
-        feat (dict): Dictionary containing the initial feature data.
-        Job (dict): Configuration dictionary for the current job.
-
-    Returns:
-        tuple: A tuple containing the modified feature dictionary and the residue list.
-    """
-
-    # Deep copy of the feature dictionary to avoid modifying the original
-    featd = copy.deepcopy(feat)
-    featd["diffusion_t"] = torch.tensor([Job["initial_t"]])
-    featd["true_frame_tensor"] = featd["true_frame_tensor"].squeeze()
-
-    # Handling residue indices if specified in Job configuration
-    residue_list = None
-    if "residue_idx" in Job:
-        residue_list = []
-        for part in Job["residue_idx"].split("/"):
-            residue_idx = part.split("-")
-            residue_indices = [int(i) for i in residue_idx]
-            residue_list.append(residue_indices)
-
-    # Generating mask for residues to be generated
-    is_gen = torch.zeros_like(feat["frame_mask"])
-    if residue_list is not None:
-        for start, end in residue_list:
-            is_gen[:, start:end + 1] = 1.
-    else:
-        is_gen = torch.ones_like(feat["frame_mask"])
-    featd["frame_gen_mask"] = feat["frame_mask"] * is_gen
-
-    return featd, residue_list
-
-def prepare_batch(featd, lab, args):
-    """
-    Prepares the batch for processing by collating feature data and labels and transferring to the specified device.
-
-    Args:
-        featd (dict): Modified feature dictionary for a specific job and replica.
-        lab (dict): Labels corresponding to the feature data.
-        args: Parsed command-line arguments containing device information.
-
-    Returns:
-        dict: The batch data ready for processing, transferred to the specified device.
-    """
-
-    # Collate the features into a batch
-    batch = DiffoldDataset.collater([featd])
-
-    # Transfer batch data to the specified device (e.g., GPU)
-    batch_device = {k: torch.as_tensor(v, device=args.device) for k, v in batch.items()}
-
-    # Include additional batch data from labels
-    chain_id_list = []
-    for chain_index in range(len(lab)):
-        chain_id = torch.tensor([chain_index + 1] * lab[chain_index]["aatype"].shape[0], device=args.device)
-        chain_id_list.append(chain_id)
-    batch_device["chain_id"] = torch.cat(chain_id_list, dim=0)
-
-    return batch_device
 
 def main(args):
     """
@@ -177,12 +86,10 @@ def main(args):
     toc = time.perf_counter() - tic
     logging.info(f"model initialized in {toc:.2f} seconds.")
     
-    print("config beta_clip", config.diffusion.position.beta_clip)
     diffuser = Diffuser(config.diffusion)
     gpu_diffuser = Diffuser(config.diffusion).to(args.device)
         
     config.data.predict.supervised = True
-    print("config global keys", config.globals.keys())
     config.globals.chunk_size = int(args.chunk_size)
     print("config chunk size", config.globals.chunk_size)
     
@@ -195,7 +102,7 @@ def main(args):
     for job_name, Job in zip(job_name_list, Job_list):
         output_dir, output_traj_dir, dir_feat_name = setup_directories(args, job_name_list, job_name, Job)
 
-        save_config_json(args, checkpoint_path, output_dir, Job)
+        utils.save_config_json(checkpoint_path, output_dir, Job)
         
         # use existing PDB and MSA datasets to get all the features and labels for query Protein ID
         if not args.from_pdb:
@@ -242,12 +149,9 @@ def main(args):
             )
             # print out the number of chains of the protein
             print("chain number", len(lab))
-            featd, residue_list = prepare_features(feat, Job)
-            print("featd frame gen mask", featd["frame_gen_mask"])
-            
+            featd, _ = utils.prepare_features(feat, Job)
             feat_list.append(featd)
             lab_list.append(lab)
-        
         # align the second stucture to the first one
         print("feat all atom pos shape", feat_list[0]["all_atom_positions"].shape)
         all_atom_pos_1 = feat_list[0]["all_atom_positions"][0].reshape(-1, 3)
@@ -273,9 +177,7 @@ def main(args):
                 logging.info(f"add noise {featd['diffusion_t']} to backbone and sidechain frame...")
                 featd = diffuse_inputs(featd, diffuser, my_seed, config.diffusion, task="predict")
 
-            batch = prepare_batch(featd, lab, args)
-            print("batch chain id shape", batch["chain_id"].shape)
-            
+            batch = utils.prepare_batch(featd, lab, args)
             job_name = job_name_list[job_name_index]
             rep_name = f"{job_name}"
             if replica == 0:
@@ -299,17 +201,9 @@ def main(args):
         if Job["inf_steps"] >= 10 else range(Job["inf_steps"]))
         time_stamps = torch.linspace(Job["initial_t"], 0., Job["inf_steps"] + 1).float().to(args.device)
             
-        print("ft_list 0 shape", ft_list[0].shape)
-        print("ft_list 1 shape", ft_list[1].shape)
         interpolate_ft_list = utils.interpolate_conf(
             ft_list[0], ft_list[1], Job["num_steps"])
-        print("interpolate ft list shape", len(interpolate_ft_list))
-        print("interpolate ft list 0 shape", interpolate_ft_list[0].shape)
         rep_name = "_".join(job_name_list)
-        if args.model == "ufconf_af2_v3_ftnx":
-            with torch.no_grad():
-                m_out, z_out, s_out = model.iteration_evoformer(batch)
-        
         # inference from interpolated noisy frames
         for step in tqdm.tqdm(range(Job["num_steps"]), total=Job["num_steps"]) \
         if Job["num_steps"] >= 10 else range(Job["num_steps"]):
@@ -319,11 +213,6 @@ def main(args):
             batch = make_noisy_quats(batch)
             s_t, f_t = batch["aatype"].squeeze(
             ), batch["noisy_frames"].squeeze()
-            # print("batch mask shape",batch["frame_gen_mask"].shape)
-            # print("batch mask", batch["frame_mask"])
-            # print("batch gen mask",batch["frame_gen_mask"])
-            # print("batch diffusion t",batch["diffusion_t"])
-            # print("batch time",batch["time_feat"])
             # output the noisy structures into pdb files
             with open(os.path.join(output_traj_dir, f"ft_inf_{rep_name}_rep{replica}.pdb"), "a") as f:
                 f.write(utils.to_pdb_string(
@@ -336,15 +225,11 @@ def main(args):
                 t, s = time_stamps[i], time_stamps[i + 1]
                 print("batch inner diffusion t", batch_inner["diffusion_t"])
                 assert batch_inner["diffusion_t"] == t
-                if args.model == "ufconf_af2_v3_ftnx":
-                    with torch.no_grad():
-                        ret_atom_pos, ret_atom_mask, ret_frames = model.run_structure_module(batch, z_out, s_out)
-                else:
-                    with torch.no_grad():
-                        out = model(batch_inner)
-                        if config.diffusion.chi.enabled:
-                            pred_chi_angles_sin_cos = out["sm"]["angles"][-1,0,:,3:,:]
-                        ret_frames = out["pred_frame_tensor"]
+                with torch.no_grad():
+                    out = model(batch_inner)
+                    if config.diffusion.chi.enabled:
+                        pred_chi_angles_sin_cos = out["sm"]["angles"][-1,0,:,3:,:]
+                    ret_frames = out["pred_frame_tensor"]
                 
                 if config.diffusion.chi.enabled:
                     batch_inner["chi_angles_sin_cos"] = batch_inner["chi_angles_sin_cos"] *  batch_inner["chi_mask"][..., None]
@@ -359,20 +244,12 @@ def main(args):
                 f_t_inner = batch_inner["noisy_frames"].squeeze()
                 
                 with torch.no_grad():
-                    if args.fwd:
-                        f_s_inner, tor_s_inner  = gpu_diffuser.addnoise(
-                            fh_0_innner,
-                            batch_inner["frame_gen_mask"],
-                            t=s,
-                            tor_s=torh_0_inner
-                        )
-                    else:
-                        f_s_inner, tor_s_inner  = gpu_diffuser.denoise(
-                            f_t_inner, fh_0_innner,
-                            batch_inner["frame_gen_mask"],
-                            t=t, s=s,
-                            tor_t=tor_t_inner, torh_0=torh_0_inner
-                        )
+                    f_s_inner, tor_s_inner  = gpu_diffuser.denoise(
+                        f_t_inner, fh_0_innner,
+                        batch_inner["frame_gen_mask"],
+                        t=t, s=s,
+                        tor_t=tor_t_inner, torh_0=torh_0_inner
+                    )
                 batch_inner["noisy_frames"] = f_s_inner
                 if tor_s_inner is not None:
                     tor_s_inner = tor_s_inner * batch_inner["chi_mask"][...,None]
@@ -398,27 +275,19 @@ def main(args):
             # last step of the prediction
             t = time_stamps[Job["inf_steps"]]
             assert batch_inner["diffusion_t"] == t
-            if args.model == "ufconf_af2_v3_ftnx":
-                with torch.no_grad():
-                    ret_atom_pos, ret_atom_mask, ret_frames = model.run_structure_module(batch, z_out, s_out)
-            else:
-                with torch.no_grad():
-                    out = model(batch_inner)
-                    ret_frames = out["pred_frame_tensor"]
-            
+            with torch.no_grad():
+                out = model(batch_inner)
+                ret_frames = out["pred_frame_tensor"]
             sh_0, fh_0 = batch_inner["aatype"].squeeze(), ret_frames.squeeze()
             b_factor = out["plddt"][..., None].tile(37).squeeze()
-            
             # output the predicted interpolation structures into pdb files
             with open(os.path.join(output_traj_dir, f"f0h_inf_{rep_name}_rep{replica}.pdb"), "a") as f:
                 f.write(utils.to_pdb_string(
                     sh_0, fh_0, **batch_constants, model_id=step + 1, b_factor=b_factor
                 ))
-
         print("Complete interpolation process!")
         
     return
-
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
@@ -426,20 +295,16 @@ if __name__ == "__main__":
     parser.add_argument(
         "-t", "--tasks", type=str, default=None, required=True,
         help=r"""
-path to a json file specifying the task to conduct.
-The json file should be organized as:
-{
-    "tid1": {"kw1": arg1, "kw2": arg2, ...},
-    "tid2": {"kw1": arg1, "kw2", arg2, ...},
-    ...
-}.
-The task id `tid` should contain only letters, digits, and `_`. 
-The keyword arguments are illustrated in ufconf/README.md.
-"""
+    path to a json file specifying the task to conduct. The keyword arguments are illustrated in ufconf/README.md.
+        """
     )
     parser.add_argument(
         "-i", "--input_pdbs", type=str, default=None,
         help="directory of input reference pdbs."
+    )
+    parser.add_argument(
+        "-msa", "--msa_file_path", type=str, default=None,
+        help="the path for existing MSA file"
     )
     parser.add_argument(
         "-o", "--outputs", type=str, default=None, required=True,
@@ -453,7 +318,6 @@ The keyword arguments are illustrated in ufconf/README.md.
         "-st", "--start", type=int, default=0,
         help="replica number start from"
     )
-
     # model arguments
     parser.add_argument(
         "--device", type=str, default="cuda:0",
@@ -472,17 +336,16 @@ The keyword arguments are illustrated in ufconf/README.md.
         help="specify random seed for MSA."
     )
     parser.add_argument(
-        "--fwd", action="store_true",
-        help="if set, then running reverse dynamics with forward process (default: not set)."
+        "--use_exist_msa", action="store_true",
+        help="if set, then use existing MSA (default: not set)."
     )
     parser.add_argument(
-        "--from_pdb", action="store_true",
-        help="if set, then running inference with MSA generated on the fly (default: not set)."
+        "--from_pdb", action=argparse.BooleanOptionalAction, default=True,
+        help="if set, then running inference with MSA generated on the fly (default: set)."
     )
     parser.add_argument(
         "--bf16", action="store_true",
         help="if set, then bfloat16 is used. (default: not set)."
     )
-
     args = parser.parse_args()
     main(args)
