+++ inference_scripts_ufconf/run_ufconf_langevin.py
@@ -5,16 +5,12 @@ import json
 import torch
 import time
 from typing import *
-import numpy as np
 import tqdm
-
 from absl import logging
 logging.set_verbosity("info")
-from ufconf.dataset import DiffoldDataset
+from unifold.dataset import process
 from ufconf.diffusion.diffuser import Diffuser
 from ufconf.diffuse_ops import diffuse_inputs, make_noisy_quats, rbf_kernel
-from unifold.data.protein import to_pdb
-from unifold.dataset import process
 import copy
 import random
 import ufconf.utils as utils
@@ -22,122 +18,7 @@ import ufconf.utils as utils
 max_retries = 3  # Maximum number of retries
 retry_delay = 3  # Delay between retries in seconds
 
-def setup_directories(args, job_name, Job):
-    """
-    Sets up and returns the paths for the main output directory, trajectory output directory, and features directory.
-
-    Args:
-        args: Parsed command-line arguments.
-        job_name (str): Name of the current job.
-        Job (dict): Configuration dictionary for the current job.
-
-    Returns:
-        tuple: A tuple containing paths to the output directory, trajectory directory, and features directory.
-    """
-
-    # Main output directory, named based on job details
-    output_dir = os.path.join(args.outputs, job_name, f"tem{Job['initial_t']}_rep{Job['num_replica']}_num{Job['num_steps']}_inf{Job['inf_steps']}")
-    os.makedirs(output_dir, exist_ok=True)
-
-    # Directory for trajectory outputs
-    output_traj_dir = os.path.join(output_dir, "traj")
-    os.makedirs(output_traj_dir, exist_ok=True)
-
-    # Directory for features
-    dir_feat_name = os.path.join(args.outputs, job_name, 'features')
-    os.makedirs(dir_feat_name, exist_ok=True)
-
-    return output_dir, output_traj_dir, dir_feat_name
-
-def save_config_json(args, checkpoint_path, output_dir, Job):
-    """
-    Saves the job configuration to a JSON file in the output directory.
-
-    Args:
-        args: Parsed command-line arguments.
-        checkpoint_path (str): Path to the model checkpoint.
-        output_dir (str): Directory where the output is to be saved.
-        Job (dict): Configuration dictionary for the current job.
-    """
-    config_data = {
-        "timestamp": time.strftime("%Y-%m-%d;%H:%M:%S", time.localtime()),
-        "diffusion_t": Job["initial_t"],
-        "checkpoint": checkpoint_path
-    }
-
-    # Define the file path
-    config_file_path = os.path.join(output_dir, "config.json")
-
-    # Writing the configuration data to the JSON file
-    with open(config_file_path, 'w') as config_file:
-        json.dump(config_data, config_file, indent=2)
-        
-def prepare_features(feat, Job):
-    """
-    Prepares features based on the job configuration.
-
-    Args:
-        feat (dict): Dictionary containing the initial feature data.
-        Job (dict): Configuration dictionary for the current job.
-
-    Returns:
-        tuple: A tuple containing the modified feature dictionary and the residue list.
-    """
-
-    # Deep copy of the feature dictionary to avoid modifying the original
-    featd = copy.deepcopy(feat)
-    featd["diffusion_t"] = torch.tensor([Job["initial_t"]])
-    featd["true_frame_tensor"] = featd["true_frame_tensor"].squeeze()
-
-    # Handling residue indices if specified in Job configuration
-    residue_list = None
-    if "residue_idx" in Job:
-        residue_list = []
-        for part in Job["residue_idx"].split("/"):
-            residue_idx = part.split("-")
-            residue_indices = [int(i) for i in residue_idx]
-            residue_list.append(residue_indices)
-
-    # Generating mask for residues to be generated
-    is_gen = torch.zeros_like(feat["frame_mask"])
-    if residue_list is not None:
-        for start, end in residue_list:
-            is_gen[:, start:end + 1] = 1.
-    else:
-        is_gen = torch.ones_like(feat["frame_mask"])
-    featd["frame_gen_mask"] = feat["frame_mask"] * is_gen
-
-    return featd, residue_list
-
-def prepare_batch(featd, lab, args):
-    """
-    Prepares the batch for processing by collating feature data and labels and transferring to the specified device.
-
-    Args:
-        featd (dict): Modified feature dictionary for a specific job and replica.
-        lab (dict): Labels corresponding to the feature data.
-        args: Parsed command-line arguments containing device information.
-
-    Returns:
-        dict: The batch data ready for processing, transferred to the specified device.
-    """
-
-    # Collate the features into a batch
-    batch = DiffoldDataset.collater([featd])
-
-    # Transfer batch data to the specified device (e.g., GPU)
-    batch_device = {k: torch.as_tensor(v, device=args.device) for k, v in batch.items()}
-
-    # Include additional batch data from labels
-    chain_id_list = []
-    for chain_index in range(len(lab)):
-        chain_id = torch.tensor([chain_index + 1] * lab[chain_index]["aatype"].shape[0], device=args.device)
-        chain_id_list.append(chain_id)
-    batch_device["chain_id"] = torch.cat(chain_id_list, dim=0)
-
-    return batch_device
-
-def process_replica(args, model, batch, diffuser, config, Job, output_traj_dir, job_name, theta_list, trans_list, replica):
+def process_replica(args, model, batch, diffuser, config, Job, output_traj_dir, job_name, replica):
     # set langevin step size
     sigma_l_r = 0.0005
     sigma_l_p = 0.01
@@ -156,33 +37,12 @@ def process_replica(args, model, batch, diffuser, config, Job, output_traj_dir, 
         os.path.join(output_traj_dir, f"f0_{rep_name}.pdb"),
         s_0, f_0, **batch_constants, b_factor=None
     )
-    # # output the ground truth structure including side chain atoms
-    # with open(os.path.join(output_traj_dir, f"f0_{rep_name}.pdb"), "w") as f:
-    #     batch_squeeze = copy.deepcopy(batch)
-    #     batch_squeeze["asym_id"] = batch["asym_id"].squeeze()
-    #     batch_squeeze["aatype"] = batch["aatype"].squeeze()
-    #     batch_squeeze["all_atom_positions"] = batch["all_atom_positions"].squeeze()
-    #     batch_squeeze["residue_index"] = batch["residue_index"].squeeze()
-    #     batch_squeeze["all_atom_mask"] = batch["all_atom_mask"].squeeze()
-    #     f.write(to_pdb(utils.make_output(batch_squeeze)))
-    
     iter_n = (tqdm.tqdm(range(Job["num_steps"]), total=Job["num_steps"]) \
         if Job["num_steps"] >= 10 else range(Job["num_steps"]))
     iter_t = (tqdm.tqdm(range(Job["inf_steps"]), total=Job["inf_steps"]) \
         if Job["inf_steps"] >= 10 else range(Job["inf_steps"]))
-    
     s_t, f_t = batch["aatype"].squeeze(), batch["noisy_frames"].squeeze()
     
-    gamma_p = diffuser.pos_trans.gamma(torch.tensor(Job["initial_t"]))
-    thetas, translations = utils.compute_theta_translation(f_t, f_0, gamma_p, batch["frame_gen_mask"])
-    
-    theta_t_list = []
-    trans_t_list = []
-    theta_t_list.append(thetas)
-    trans_t_list.append(translations)
-    if args.model == "ufconf_af2_v3_ftnx":
-        with torch.no_grad():
-            m_out, z_out, s_out = model.iteration_evoformer(batch)
     for n_step in iter_n:
         # save noisy frames 
         if Job["save_trajectory"]:
@@ -199,15 +59,11 @@ def process_replica(args, model, batch, diffuser, config, Job, output_traj_dir, 
         for i in iter_t:
             t, s = time_stamps[i], time_stamps[i + 1]
             assert batch_inner["diffusion_t"] == t
-            if args.model == "ufconf_af2_v3_ftnx":
-                with torch.no_grad():
-                    ret_atom_pos, ret_atom_mask, ret_frames = model.run_structure_module(batch, z_out, s_out)
-            else:
-                with torch.no_grad():
-                    out = model(batch_inner)
-                    if config.diffusion.chi.enabled:
-                        pred_chi_angles_sin_cos = out["sm"]["angles"][-1,0,:,3:,:]
-                    ret_frames = out["pred_frame_tensor"]
+            with torch.no_grad():
+                out = model(batch_inner)
+                if config.diffusion.chi.enabled:
+                    pred_chi_angles_sin_cos = out["sm"]["angles"][-1,0,:,3:,:]
+                ret_frames = out["pred_frame_tensor"]
             # aligne X0_h with X_t
             batch["noisy_frames"][0], out["pred_frame_tensor"] = utils.remove_center(batch["noisy_frames"][0], out["pred_frame_tensor"], mask=batch["frame_mask"])
             noisy_pos, noisy_mask = utils.compute_atomic_positions(batch["noisy_frames"], **batch_constants)
@@ -218,7 +74,7 @@ def process_replica(args, model, batch, diffuser, config, Job, output_traj_dir, 
             predicted_pos, predicted_mask = utils.compute_atomic_positions(out["pred_frame_tensor"], **batch_constants)
             predicted_pos = predicted_pos[0].reshape(-1,3)
             
-            prediced_pos_aligned, U = utils.kabsch_rotate(predicted_pos.cpu().numpy(), noisy_pos.cpu().numpy())
+            prediced_pos_aligned, U = utils.kabsch_rotate(predicted_pos, noisy_pos)
             prediced_pos_aligned = prediced_pos_aligned.reshape(1, batch["aatype"].shape[-1], -1, 3)
             batch["all_atom_positions"] = torch.tensor(prediced_pos_aligned,device = args.device)
             out["final_atom_positions"] = batch["all_atom_positions"] 
@@ -238,20 +94,12 @@ def process_replica(args, model, batch, diffuser, config, Job, output_traj_dir, 
             f_t_inner = batch_inner["noisy_frames"].squeeze()
             
             with torch.no_grad():
-                if args.fwd:
-                    f_s_inner, tor_s_inner  = diffuser.addnoise(
-                        fh_0_innner,
-                        batch_inner["frame_gen_mask"],
-                        t=s,
-                        tor_s=torh_0_inner
-                    )
-                else:
-                    f_s_inner, tor_s_inner  = diffuser.denoise(
-                        f_t_inner, fh_0_innner,
-                        batch_inner["frame_gen_mask"],
-                        t=t, s=s,
-                        tor_t=tor_t_inner, torh_0=torh_0_inner
-                    )
+                f_s_inner, tor_s_inner  = diffuser.denoise(
+                    f_t_inner, fh_0_innner,
+                    batch_inner["frame_gen_mask"],
+                    t=t, s=s,
+                    tor_t=tor_t_inner, torh_0=torh_0_inner
+                )
             batch_inner["noisy_frames"] = f_s_inner
             if tor_s_inner is not None:
                 tor_s_inner = tor_s_inner * batch_inner["chi_mask"][...,None]
@@ -285,75 +133,32 @@ def process_replica(args, model, batch, diffuser, config, Job, output_traj_dir, 
             )
             time_feat = rbf_kernel(residue_t, config.diffusion.d_time, 0., 1.)
             batch_inner["time_feat"] = time_feat
-            
         # last step of the prediction
         t = time_stamps[Job["inf_steps"]]
         assert batch_inner["diffusion_t"] == t
-        if args.model == "ufconf_af2_v3_ftnx":
-            with torch.no_grad():
-                ret_atom_pos, ret_atom_mask, ret_frames = model.run_structure_module(batch, z_out, s_out)
-        else:
-            with torch.no_grad():
-                out = model(batch_inner)
-                ret_frames = out["pred_frame_tensor"]
-        
+        with torch.no_grad():
+            out = model(batch_inner)
+            ret_frames = out["pred_frame_tensor"]
         sh_0, fh_0 = batch_inner["aatype"].squeeze(), ret_frames.squeeze()
-
         if Job["save_trajectory"]:
             # output the predicted structures during reverse process into pdb files
             with open(os.path.join(output_traj_dir, f"f0h_inf_{rep_name}_num{n_step}.pdb"), "a") as f:
                 f.write(utils.to_pdb_string(
                     sh_0, fh_0, **batch_constants, model_id=Job["inf_steps"] + 1,
                 ))
-        
         # output the predicted structure with only backbones
         with open(os.path.join(output_traj_dir, f"f0h_inf_{rep_name}_rep{replica}.pdb"), "a") as f:
             f.write(utils.to_pdb_string(
                     sh_0, fh_0, **batch_constants, model_id=n_step,
                 ))
-            # output the predicted structure including side chain atoms
-            # batch_squeeze = copy.deepcopy(batch_inner)
-            # out_squeeze = copy.deepcopy(out)
-            # batch_squeeze["asym_id"] = batch_inner["asym_id"].squeeze()
-            # batch_squeeze["aatype"] = batch_inner["aatype"].squeeze()
-            # batch_squeeze["residue_index"] = batch_inner["residue_index"].squeeze()
-            # out_squeeze["final_atom_positions"] = out["final_atom_positions"].squeeze()
-            # out_squeeze["final_atom_mask"] = out["final_atom_mask"].squeeze()
-            # prot = utils.make_output(batch_squeeze, out_squeeze)
-            # f.write(to_pdb(prot, model_id=n_step + 1))
-        
         f_s = diffuser.langevin(
             f_t, fh_0, batch["frame_gen_mask"],
-            t=batch["diffusion_t"].squeeze(), sigma_l_r=torch.tensor(sigma_l_r, device="cuda:0"), sigma_l_p=torch.tensor(sigma_l_p)
+            t=batch["diffusion_t"].squeeze(), sigma_l_r=torch.tensor(sigma_l_r, device=args.device), sigma_l_p=torch.tensor(sigma_l_p)
         )
-            
         batch["noisy_frames"] = f_s
         s_t, f_t = batch["aatype"].squeeze(), batch["noisy_frames"].squeeze()
-    
-        gamma_p = diffuser.pos_trans.gamma(torch.tensor(batch["diffusion_t"].squeeze()))
-        thetas, translations = utils.compute_theta_translation(f_t, f_0, gamma_p, batch["frame_gen_mask"])
-        theta_t_list.append(thetas)
-        trans_t_list.append(translations)
-    
         # update features
         batch = make_noisy_quats(batch)
-    
-    # save final
-    # with open(os.path.join(output_dir,f"f0h_{rep_name}.pdb"), "a") as f:
-    #     f.write(to_pdb_string(
-    #         sh_0, fh_0, **batch_constants, model_id=replica+1
-    #     ))
-    theta_list.append(theta_t_list)
-    trans_list.append(trans_t_list)
-    
-def run_job(args, model, batch, gpu_diffuser, config, Job, output_traj_dir, job_name, \
-            theta_list, trans_list, replica, available_gpus):
-    gpu_id = available_gpus.pop(0)  # Allocate GPU
-    os.environ['CUDA_VISIBLE_DEVICES'] = str(gpu_id)
-    result = process_replica(args, model, batch, gpu_diffuser, config, Job, output_traj_dir, job_name, \
-            theta_list, trans_list, replica)
-    available_gpus.append(gpu_id)  # Release GPU
-    return result
 
 def main(args):
     """
@@ -393,12 +198,10 @@ def main(args):
     toc = time.perf_counter() - tic
     logging.info(f"model initialized in {toc:.2f} seconds.")
     
-    print("config beta_clip", config.diffusion.position.beta_clip)
     diffuser = Diffuser(config.diffusion)
     gpu_diffuser = Diffuser(config.diffusion).to(args.device)
         
     config.data.predict.supervised = True
-    print("config global keys", config.globals.keys())
     config.globals.chunk_size = int(args.chunk_size)
     print("config chunk size", config.globals.chunk_size)
     
@@ -407,10 +210,8 @@ def main(args):
 
     # Iterate over each job
     for job_name, Job in zip(job_name_list, Job_list):
-        output_dir, output_traj_dir, dir_feat_name = setup_directories(args, job_name, Job)
-
-        save_config_json(args, checkpoint_path, output_dir, Job)
-        
+        output_dir, output_traj_dir, dir_feat_name = utils.setup_directories(args, job_name, Job, mode = "langevin")
+        utils.save_config_json(checkpoint_path, output_dir, Job)
         # use existing PDB and MSA datasets to get all the features and labels for query Protein ID
         if not args.from_pdb:
             print(f"Please provide valid flag: 'from_pdb'.")
@@ -431,15 +232,11 @@ def main(args):
                         print("Maximum retries reached. Aborting.")
                         raise
                     
-        theta_list = []
-        trans_list = []
         for replica in tqdm.tqdm(range(Job["num_replica"]), total=Job["num_replica"]) \
             if Job["num_replica"] >= 10 else range(Job["num_replica"]):
             logging.info(f"running replica {replica+1}/{Job['num_replica']}...")
             replica = replica + args.start * Job["num_replica"]
-            
             my_seed = random.randint(0, 1000000)  # Generate a random seed
-            
             # preprocess all the MSA, make random deletion to the MSAs controlled by the random seed defined in `data_idx`
             feat, lab = process(
                 config.data,
@@ -450,14 +247,10 @@ def main(args):
                 data_idx=my_seed,
                 is_distillation=False
             )
-            
             # print out the number of chains of the protein
             print("chain number", len(lab))
-            
-            featd, residue_list = prepare_features(feat, Job)
-            print("featd frame gen mask", featd["frame_gen_mask"])
-
-            # my_seed = 10
+            featd, _ = utils.prepare_features(feat, Job)
+            my_seed = random.randint(0, 1000000)  # Generate a random seed
             # diffuse inputs to noisy structure
             if featd["diffusion_t"] == 1.0:
                 logging.info(f"set backbone and sidechain frame to be prior...")
@@ -465,36 +258,11 @@ def main(args):
             else:
                 logging.info(f"add noise {featd['diffusion_t']} to backbone and sidechain frame...")
                 featd = diffuse_inputs(featd, diffuser, my_seed, config.diffusion, task="predict")
-
-            batch = prepare_batch(featd, lab, args)
-            print("batch chain id shape", batch["chain_id"].shape)
-            
-            process_replica(args, model, batch, gpu_diffuser, config, Job, output_traj_dir, job_name, \
-                theta_list, trans_list, replica)
-            
+            batch = utils.prepare_batch(featd, lab, args)
+            process_replica(args, model, batch, gpu_diffuser, config, Job, output_traj_dir, job_name, replica)
         print("Langevin inference completed!")
-        theta_list = np.array(theta_list)
-        trans_list = np.array(trans_list)
-        print("theta_list", theta_list.shape)
-        print("trans_list", trans_list.shape)
         
-        theta_all_reshaped = theta_list.transpose(1, 0, 2).reshape(Job["num_steps"] + 1, -1)
-        trans_all_reshaped = trans_list.transpose(1, 0, 2, 3, 4).reshape(Job["num_steps"] + 1, -1)
-        np.savetxt(os.path.join(output_dir, "theta_all.txt"), theta_all_reshaped)
-        np.savetxt(os.path.join(output_dir, "trans_all.txt"), trans_all_reshaped)
-        
-        for time_index in range(0, Job["num_steps"], Job["num_steps"] // 5):
-            theta_time_list = theta_list[:, time_index]
-            pos_time_list = trans_list[:, time_index]
-
-            flattened_theta_time = theta_time_list.flatten()
-            flattened_pos_time = pos_time_list.flatten()
-
-            np.savetxt(os.path.join(output_dir, f"theta{time_index}.txt"), flattened_theta_time)
-            np.savetxt(os.path.join(output_dir, f"trans{time_index}.txt"), flattened_pos_time)
-            
     return
-
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
@@ -502,20 +270,16 @@ if __name__ == "__main__":
     parser.add_argument(
         "-t", "--tasks", type=str, default=None, required=True,
         help=r"""
-path to a json file specifying the task to conduct.
-The json file should be organized as:
-{
-    "tid1": {"kw1": arg1, "kw2": arg2, ...},
-    "tid2": {"kw1": arg1, "kw2", arg2, ...},
-    ...
-}.
-The task id `tid` should contain only letters, digits, and `_`. 
-The keyword arguments are illustrated in ufconf/README.md.
-"""
+    path to a json file specifying the task to conduct. The keyword arguments are illustrated in ufconf/README.md.
+        """
     )
     parser.add_argument(
         "-i", "--input_pdbs", type=str, default=None,
         help="directory of input reference pdbs."
+    )
+    parser.add_argument(
+        "-msa", "--msa_file_path", type=str, default=None,
+        help="the path for existing MSA file"
     )
     parser.add_argument(
         "-o", "--outputs", type=str, default=None, required=True,
@@ -529,7 +293,6 @@ The keyword arguments are illustrated in ufconf/README.md.
         "-st", "--start", type=int, default=0,
         help="replica number start from"
     )
-
     # model arguments
     parser.add_argument(
         "--device", type=str, default="cuda:0",
@@ -548,17 +311,16 @@ The keyword arguments are illustrated in ufconf/README.md.
         help="specify random seed for MSA."
     )
     parser.add_argument(
-        "--fwd", action="store_true",
-        help="if set, then running reverse dynamics with forward process (default: not set)."
+        "--use_exist_msa", action="store_true",
+        help="if set, then use existing MSA (default: not set)."
     )
     parser.add_argument(
-        "--from_pdb", action="store_true",
-        help="if set, then running inference with MSA generated on the fly (default: not set)."
+        "--from_pdb", action=argparse.BooleanOptionalAction, default=True,
+        help="if set, then running inference with MSA generated on the fly (default: set)."
     )
     parser.add_argument(
         "--bf16", action="store_true",
         help="if set, then bfloat16 is used. (default: not set)."
     )
-
     args = parser.parse_args()
     main(args)
